tweeti/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Django==3.1.3
django-filter==2.4.0
djangorestframework==3.12.4
filelock==3.0.12
idna==2.10
importlib-metadata==4.5.0
lxml==4.6.2
Markdown==3.3.4
oauthlib==3.1.0
pandas==1.2.0
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.5
requests==2.25.1
requests-file==1.5.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
tldextract==3.1.0
tqdm==4.55.2
tweepy==3.10.0
typing-extensions==3.10.0.0
unshortenit==0.4.0
urlexpander==0.0.36
urllib3==1.26.2
zipp==3.4.1
django-anymail==8.6
gunicorn==20.1.0
psycopg2-binary==2.9.1
